Update build scripts to support code snippets other than 'C'
Add code to fcvextract.pl to switch regex depending on the suffix
of source file. Supported suffixes are ".c", ".h", ".sh", ".spin",
and ".litmus".

Also fix gen_snippet_d.pl so that it works when a source file under
a deep sub-directory is specified.

Some of the <file name> part specified in the 2nd arg to fcvextract.pl
requires escaping. Specifically, "+" characters e.g. in "C-SB+o-o+o-o"
need escapes and should be specified as "C-SB\\+o-o\\+o-o" in the
argument. This escaping can be most easily done in Makefile by
using the $(subst ...) function.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>