Add scripts and recipes to work around restriction of herdtools7
It turned out that .litmus files do not have so much freedom
in placement of comments, and what words can appear in comments.

To work around the restriction, we need to add another route to
generate VerbatimL LaTeX sources (.fcv) from litmus tests (.litmus).

    .litmus -> .ltms
    .ltms   -> .fcv

In .litmus, we place a meta command to start a code snippet below
the line of "C foo-bar", which needs to be the first line in .litmus
file.

Also, as the "filter" and "exists" clause don't permit any comments
following the, we place a meta command to end a code snippet ahead
of the line of "exists" or "filter" clause.
And "{" and "}" in these meta commands sometimes cause error depending
on their positions. So we use the form of:

    \begin[snippet][option] -- \end[snippet]

in .litmus files.

For consistency, \lnlbl commands should be of the form of \lnlbl[bar].

NOTE:
    Comments of the form  "(* ... *)"  can not appear in C-language
    parts of litmus tests. The form of "//\lnlbl{bar}" is OK in
    C-language parts.

The reordering of these meta commands is handled by the new
script, utilities/reorder_ltms.pl. It also takes care of
"[" and "]" characters.

In summary, .litmus source should look like the following:

------------------------------------------------
C C-foo-bar
//\begin[snippet][labelbase=foo:bar,...]

{
}

[...]

    smp_mb(); //\lnlbl[fullbarrier]

[...]
}

//\end[snippet]
exists (1:r2=0 /\ 0:r2=0)
------------------------------------------------

This will result in a .ltms file:

------------------------------------------------
//\begin{snippet}[labelbase=foo:bar,...]
C C+foo-bar

{
}

[...]

    smp_mb(); //\lnlbl{fullbarrier}

[...]
}

exists (1:r2=0 /\ 0:r2=0)
//\end{snippet}

Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>