reorder_ltms.pl: Enable further labeling in litmus test snippet
In .litmus files, lines of "filter" and "exists" clauses can't
have their own comments to put labels.
To overcome this restriction, we add options "existslabel=foo"
and "filterlabel=bar" to the \end[snippet] meta command.

To put a label to an exists clause, you can put the label in
the form:

    ...
    \end[snippet][existslabel=foo]
    exists (1:r2=0 /\ 0:r2=0)
    [EOF]

This will be converted to:

    ...
    exists (1:r2=0 /\ 0:r2=0)\lnlbl{foo}
    \end{snippet}
    [EOF]

in .ltms file.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>