future/formalregress: Example of extraction of snippet from .litmus file
This is a POC of including litmus test snippet extracted via
reorder_ltms.pl and fcvextract.pl.

As herdtools7 does not allow comment following a "filter" or
"exists" clause, we use an option "filterlabel" to the meta command
"\end[snippet]".

NOTE: Neither "exists" nor "filter" can appear as a token in
    comments of litmus test. "filter_" is OK. A token in comment?
    Don't ask me!

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>