count: Fix uses of READ/WRITE_ONCE()s in count_end and count_tstat
Add/remove READ_ONCE()/WRITE_ONCE()s to comply with the
guideline to be expanded in toolsoftrade.

Note on addition of READ_ONCE() while holding a lock:
In read_count() of count_end.c, READ_ONCE() is added to access
of *counterp[t] while holding final_mutex. This is because
final_mutex protects only counterp[], but per-thread variables
pointed to by the array are updated by the owning threads'
inc_count().

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>