atomic_cmpxchg: Handle spurious failure of __atomic_compare_exchange_n
If the 4th parameter (field *weak*) of __atomic_compare_exchange_n is
set, it may refuse to perform the CAS operation and return a failure,
even if the value of field *old* is equal to the content of the specified
memory location. This happens in architectures where LL/SC is used to
emulate CAS primitive (e.g., PPC and ARM).

This patch returns the *old* value if __atomic_compare_exchange_n has
been successfully performed; otherwise, *old*+1 is returned, which forces
the caller to retry the CAS loop.

Signed-off-by: Junchang Wang <junchangwang@gmail.com>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>