memorder: Add a brief description of the use of multicopy atomicity
Subsection Multicopy Atomicity is ahead of Subsection Hardware Specifics where
implementation detail of multicopy atomicity in mainstream architectures is
presented. So we add (1) a brief description of the use of multicopy atomicity
in representative architectures and (2) a reference to Table Summary of Memory
Ordering, in subsection Multicopy Atomicity.

The description is borrowed from a nice discussion with Paul, which can be found
at https://www.spinics.net/lists/perfbook/msg01952.html.

Signed-off-by: Junchang Wang <junchangwang@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>