memorder: Clean up reference to lockless_dereference()
The lockless_dereference() primitive has been obsolete since Linux
kernel version v4.15, but it is mentioned in several places in the
memory-ordering chapter and in a litmus test.  This commit removes that
litmus test, removes the definition of lockless_dereference(), makes
READ_ONCE() work properly on DEC Alpha, and clearly labels remaining
instances of lockless_dereference() in the text as obsolete.

Reported-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>