SMPdesign: Use ceiling symbol, skip odd floor computation
Equation E.6 in Quick Quiz 6.20 computes the floor function of
"(m + s - 1) / s", which is just the ceiling function of "m / s".
This commit therefore moves to this simpler formulation, on the
strength of the ceil() function in the standard library.  The fact
that ceil() is not globally available in the Linux kernel, which
motivated the original, seems a bit irrelevant, hence the change.

Reported-by: Mark Morrissey <markem@pdx.edu>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>