datastruct/hash: Unlock what is locked during resize
Currently, hashtab_unlock_mod() is passed a key, which it must look up
in order to find the bucket containing the lock to be released.  This
is wasteful due to calculating the hash twice, and also accident-prone
when running concurrently with a resize operation.  This commit therefore
has hashtab_lock_mod() (which is now resize_lock_mod()) record the bucket
locked, and this record is now passed to hashtab_unlock_mod() (which is
now resize_unlock_mod()).

In addition, the original hashtab_lock_mod() acquires the old hash
table's bucket's ->lock, which prevents resizing from passing through
that bucket.  However, if this bucket has already been resized, it drops
the lock immediately after acquiring that of the new hash table, which
permits the resize to pass through the current bucket.  This means that
hashtab_unlock_mod() might not see the same state, and thus cannot know
which lock to drop.

This commit therefore also retains the old hash table's bucket's ->lock
throughout the operation, so that hashtab_unlock_mod() can know whether
to drop the old table's lock on the one hand or both on the other.
And this commit also hashtab_unlock_mod() to handle these locks correctly.

Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>