datastruct/hash: Fill in the ->hte_hash field for BUG_ON()
The hashtorture.h tests check the value of ->hte_hash, but hash_resize.c
doesn't set it, resulting in deterministic failure.  This commit therefore
fills this field in.

This approach does result in a duplicate hash-function evaluation in
hashtab_add(), which is currently left as an exercise for the reader.

Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>