datastruct/hash: Debug checks prior to hashtab_add() revamp
The goal is to make hashtab_add() refrain from rehashing and to make
the coupling of the locking and the actual linked-list manipulation
be correct by construction.  But for that to be the case, it is necessary
for the values computed at hashtab_lock_mod() time to actually match
those computed at hashtab_add() time.

This commit adds debug checks to check this matching.

Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>