datastruct/hash: Publish new hash table after initializing it
Currently, when resizing, the new hash table is published before the
->ht_idx field is set.  This is in theory OK because the new hash
table cannot actually be accessed until after the first update of
->ht_resize_cur (which is preceded by a full memory barrier), but in
practice this is an accident waiting to happen.

This commit therefore makes the publication be smp_store_release(),
moves it after the ->ht_idx update, and also moves the synchronize_rcu()
to follow that.  The synchronize_rcu() move is required to ensure that
readers are aware of the new hash table before the actual resizing starts.

Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>