datastruct/hash: Simplify hashtab_lookup()
Because resizing leaves the old hash table intact, and because lookups
are carried out within RCU read-side critical sections (which prevent
a second resizing operation from starting), there is no need for a
lookup to search anywhere but in the old hash table.  And in the common
case, there is no resize, so there is no new hash table.  Therefore,
eliminating the check for resizing speeds things up in the common
case.  In addition, this simplifies the code.

This commit therefore eliminates the ht_get_bucket() function,
renames the ht_get_bucket_single() function to ht_get_bucket(),
and modifies callers appropriately.

Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>