datastruct/hash: Remove extraneous barrier from hashtab_resize()
Now that hashtab_lookup() is iresizing-agnostic, all non-initialization
accesses to ->ht_resize-cur are protected by locking in the restricted
sense that any change to ->ht_resize_cur that would change the value
of the "if" condition cannot happen while the lock is held on the old
bucket.  This means that the memory barrier may be removed.  However,
the READ_ONCE() and WRITE_ONCE() markings on non-initialization accesses
to ->ht_resize_cur must remain because reads from ->ht_resize_cur really
can race with writes, just not is a way to change the "if" conditions.

Reported-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>