datastruct/hash: Make hashtab_lookup() more responsive during resize
If hashtab_add() adds a new element to an already-resized bucket during
a resize, hashtab_lookup() won't find it until the resize is complete.
This is a not-unreasonable semantic, but might be quite surprising to most
users, especially if the thread doing the lookup is the same thread that
just added the new element.  This commit therefore causes hashtab_lookup()
to find recently added elements even during resize operations.

Note that this change involved a small refactoring of the code,
introducing a new ht_search_bucket() helper function.  This refactoring
avoids the need for memory barriers as well because when an element is
not found and there is a resize in progress, the new version of the
hash table is unconditionally searched, thus avoiding the need for a
racy access to ->ht_resize_cur.  Although this approach can result in
needless searches of not-yet-filled-in buckets, such searches will
be rare (assuming resizing is rare), and a search of an empty bucket
is cheap anyway.

Reported-by: Junchang Wang <junchangwang@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>