datastruct/hash: Make hashtab_del() more thorough
During a resize, if an element is deleted after it has been inserted
into the new version of the hash table, a subsequent lookup will still
find that element, even when it is looked up by the same thread that
deleted this.

This commit therefore makes hashtab_del() remove the element from
both versions of the hash table.

Reported-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>