formal/spinhint: Update tables of memory usage of Spin
Update tables of memory usage of increment.spin and qrcu.spin.

As for qrcu.spin, to complete search of 3 updaters and 2 readers
on a machine with 16GB memory, the "-DCOLLAPSE" flag is necessary
to avoid memory exhaustion. This doesn't affect the completeness of
state search. Add explantion of the flag and update the table of
qrcu.spin with the usage obtained with the flag enabled.

There seems to have been incomplete searches due to insufficient
search depth.  Update "# state" figures from the complete results
obtained with deeper search limits.

Depth reached:
    updaters readers   depth
    1        1            95
    1        2           218
    1        3           385
    2        1           859
    2        2          2352
    2        3         12857
    3        1         53809
    3        2        328014

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>