formal/spinhint: Place footnote inside floating table
It turns out the "savenotes" environment tried in commit
5df9ce3b9449 ("formal/spinhint: Put footnote on header in table")
is not robust enough. Depending on the layout, text of a footnote
can appear at the bottom of a page ahead of (or behind) the page
where the floating table itself is typeset.

Putting such footnotes inside a floating table by the help of the
"threeparttable" package looks like an orthodox way recommended
in the LaTeX community [1].

[1]: https://texfaq.org/FAQ-footintab

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>