formal/spinhint: Update memory usage in QRCU Spin summary table
The memory usage and runtime data of 3 readers and 3 updaters
presented in Table 12.3 were from a run with a depth limit an
order of magnitude larger than the reached depth. Update the data
from a just finished run with a tighter depth limit.

Also fix a typo and substitute "Spin" for "spin" for consistency.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>