Suppress warning from pdfTeX on figures transformed by Inkscape
This change suppresses warnings such as:

    <./cartoons/r-2014-branch-error.pdf

    pdfTeX warning: pdflatex (file ./cartoons/r-2014-branch-error.pdf):
    PDF inclusion: multiple pdfs with page group included in a single page>

This is a known issue of pdfTeX [1].
The message is emitted when multiple PDF files generated by
Inkscape appear in a single page. This issue has not caused
any problem in perfbook so far.

The variable \pdfsuppresswarningpagegroup is available since
pdfTeX 1.40.15 (TeX Live 2014). So this change doesn't affect
TeX Live 2013/Debian on Ubuntu Trusty.

[1]: https://tex.stackexchange.com/questions/183149/

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>