Makefile: Add optional recipe to detect conflict in page group object
As long as "page group objects" are identical, setting
\pdfsuppresswarningpagegroup variable done in the previous commit
is safe.

This commit adds an optional recipe in ".svn -> .pdf"
transformation to extract page group object from the generated .pdf
file, and to compare it with the expected one given in "pagegroup".

The added recipe is enabled only when an environment variable
"PERFBOOK_CHKPAGEGROUP" is set "on". The check requires a command
named "qpdf", which is available on most Linux distributions.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.ibm.com>