Make hyperlinks colorful and browsable in PDF viewer
Balbir suggested coloring hyperlinks to make them easier to
recognize [1].

As 2c layouts are supposed to be for printing, enable coloring
of hyperlinks only in 1c layout. To do this, add a boolean
variable "colorlinks" and set it true for 1c layout. This would
make it easy to change the condition to enable "colorlinks".

Select "MediumBlue" via xcolor's "svgnames" option as it looks
less distracting than the pure "blue", but still distinguishable.

Redundant framing of links in bibliography is disabled when
"colorlinks" is effective.

Also make the page numbers in TOC hyperlinks as well.

[1]: https://www.spinics.net/lists/perfbook/msg02350.html

Suggested-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>