Specify required version of 'titlesec' package
It turned out that "titlesec" package introduced in commit
9bac0c81a7b4 ("Fix cross reference to paragraph") has older
versions which are incompatible with current perfbook codebase.

By specifying required version of the package in the preamble,
make it easier for people who might still keep old TeX Live
installations and end up in build error of perfbook to find
out the incompatible package.

Notes:
    TeX Live 2015/Debian on Ubuntu 16.04 has picked up a buggy
    version of titlesec [2016/03/15]. The bug was fixed shortly
    after [2016/03/21], which was picked up by Ubuntu 18.04.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>