datastruct: Clarify caption and discussion of Figure 10.17
The figure comparing the overhead of fixed-size to resizable hash tables
is a bit confusing.  It is not immediately clear what the numbers in
the diagram refer to, and it is all too natural to assume that they
refer to the number of buckets.  In fact, they refer to the number of
elements in hash tables that have either 1024 or 2048 buckets.  This
commit therefore clarifies the caption and the discussion.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>