bib: Drag format kicking and screaming into the 21st century
The .bib files use a format dating from the 1980s, when the comma was
always a separator and never a terminator.  This made editing .bib files
quite annoying due to the last line having to omit the trailing comma.
The workaround was to use comma as a prefix, so that each line starts
with comma, working around the annoying last-line special case.

Unfortunately, modern .bib-checking tools are not amused by the leading
comma, so this commit moves the commas to ends of lines.

Suggested-by: Stamatis Karnouskos <karnouskos@ieee.org>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>