utilities/autodate: Mark released versions
If there is a tag on the current commit, then that tag is the name of
the release.  This will change to accommodate editions Real Soon Now.
Current guess is to make the edition name be English: "Second Edition"
and "Second Print Edition", then add code to make any tag that doesn't
match "vYYYY.MM.DDX" to be output verbatim, without the initial "Release".

I replaced the "env printf" with an "echo" because the former did very
strange things.  Please feel free to inform me of the error of my "env
printf" ways!

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>