utilities/autodate: Put commit-id in footer by 'draftwatermark'
A similar footer should be able to be typeset on every page by
a LaTeX package "fancyhdr", but I have not figured out a way to do
so reliably on Chapter pages.

A Release/Edition will have the release/edition tag name instead.

Also update FAQ-BUILD.txt to mention the possible version mismatch
of draftwatermark.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>