Prevent section epigraph from orphaned
Latest "epigraph" package has a nice feature to make sure that
section epigraph is not orphaned.

Add version info to its \usepackage.
Also add it to #10 in FAQ-BUILD.txt.

Also suppress widows and orphans by using the "nowidow" package.
This prevents a single line of a paragraph to appear at the
bottom or the top of a page/column. For example, Short QQZ (less
than 4 lines) won't be broken in the middle any more.

This narrows the freedom of LaTeX's typesetting and slightly
increases page count, but should improve the ease of read.

NOTE 1: epigraph's document now mentions a macro \epigraphnoindent
which would have the same effect as \NoIndentAfterCmd{\epigraph},
but it has weird side effect on the spacing of other section
headings.

NOTE 2: This update causes stretched vertical spaces between
paragraphs in pages where section headings with epigraphs are
forwarded to the next pages.  They might need tweaks in LaTeX
sources later, especially in the -RC stages of the next edition.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>