count: Promote code snippet in Quiz part of QQZ to listing
Current scheme of QQZ can't handle a snippet at the end of a Quiz
part due to the implied break-ability in front of the ending "square
mark".
QQZ 5.29 suffers from such a symptom.

Promote the snippet to a proper listing and rephrase the Quiz
and Answer part accordingly.

Also add WRITE_ONCE() and adjust indent in the snippet for
consistency with the right snippet.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>