utilities/fixsvgfonts-urwps.sh: Use symbol font available on F31
Fedora 31 upgraded pango to 1.44 which abandoned support of bitmap
and type1 fonts such as StandardSymbolsPS.t1 [1].
This change affected some of .svg figures under advsync/.
Use MdSymbol (.otf) as an alternative which should be available in
texlive-collection-fontsextra on Fedora.
Or it can be installed on its own as texlive-mdsymbol.

Update Makefile and FAQ-BUILD.txt to add existence check of MdSymbol.

Also remove a few unnecessary rules for symbol fonts in scripts for
.eps files and Ubuntu's .svg files.

[1]: https://blogs.gnome.org/mclasen/2019/05/25/pango-future-directions/

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>