Rename environments 'linelabel' and 'lineref'
It turns out that the "lineno" package has variables \linelabel
and \lineref as its parameters.
The "fvextra" package, which enhances fancyvrb and enables (semi-)
automatic line breaking in code snippets, requires the "lineno"
package.

This commit renames our custom environments of the same name to
"fcvlabel" and "fcvref" for us to be able to make use of fvextra.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>