Remove required version of 'epigraph'
It turned out that the up-to-date "epigraph" doesn't work well
with the "nowidow" package on TeX Live 2015/Debian (Ubuntu Xenial)
and TeX Live 2016.

Instead, suggest upgrade when the old version is detected and
TeX Live version is compatible with the up-to-date "epigraph"
package.

This permits TeX Live 2015/Debian (Ubuntu Xenial) to complete
the build with the older "epigraph".

Update FAQ-BUILD.txt accordingly.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>