Makefile: grep only .tex files in 'linelabel' and 'lineref' search
For the checks added in commit 60c2079f2765 ("Makefile: Check
'linelabel' and 'lineref' used as environment"), it should be
sufficient to grep files listed in $(LATEXSOURCES) only.

Using this variable rather than the grep option "-R" will reduce
the cost of the checks.  We can also get rid of "patsubst" and
"filter-out" functions in the following macro definitions.
Furthermore, as the files in $(LATEXSOURCES) are not modified
during the build, we can use the operator ":=" there.

Also reorder and rephrase error messages in case output of grep
gets long.

Add .c source file(s) under appendix/styleguide/ to the definition
of "LST_SOURCES" to improve coverage of dependencies as well.

In the definitions of "SOURCES_OF_SNIPPET_ALL" and "SOURCES_OF_LITMUS",
replace the grep option "-R" (recursive option to follow symbolic
links) with the option "-r", which should be good enough.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>