Permit occasional orphans
Recent update of upstream LaTeX caused excessively wide spaces between
paragraphs in 2c layout similar to the issue mentioned in the change
log of commit 4c6ee39761a0 ("Remove required version of 'epigraph'")
at the beginning of Chapter 6.

The root cause of the unstable behavior is the constraints imposed
by the unfortunate combination of a fairly tall unbreakable block
(consisting of a section heading, a section epigraph, a paragraph
of only 3 lines), a footnote, and a floating figure.

Note that the "nowidow" package requires at least 4 lines for a
paragraph to be broken.

An orphan line (first line of a paragraph at the bottom of a page/column)
looks less ugly than a widow line (final line of a paragraph at the top
of a page/column) does.

As a band-aid patch, this commit permits an orphan at the problematic
paragraph by the \noclub command provided by the "nowidow" package
and avoids such ugly wide spaces.
Another approach would be to add some words to the paragraph and
make it 4 lines or more.

By this change, TeX Live 2015/Debian (on Ubuntu Xenial) with the
up-to-date "epigraph" can also produce the (mostly) same output.

Do similar tweaks to avoid wide vertical spaces in a couple of other
chapters.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>