SMPdesign: Fix meaningless consecutive blank pages
In the introduction of Dining Philosophers, there is two-page
"think about it" gap before presenting an answer.

Let's see what we need for print editions.

When the question is on an odd page, the answer can be on the next page
(the other side of the question's page).
When the question is on an even page, the answer should be on the
next even page to prevent it from being spotted too early.

For electronic editions, there needs to be a gap of at least one
page regardless of the question's position.

As \cleardoublepage changes its behavior depending on whether it is
on an odd page (one blank page) or on an even page (nothing), it
can not cover the requirement for electronic editions.

For our purpose, the sequence of \clearpage, "((Intentional blank page)",
and another \clearpage should suffice.

Reported-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>