Makefile: Add check of \crefrange and \clnrefrange misuse
To prevent typos simillar to the one fixed in commit b3b06a98bb7c
("datastruct.tex: fix some minor typos") to happen again,
add checks to detect the following patterns of misuses:

... \cref{label A}{label B} ...

... \cref{long label X}
{another long label Y}

In the formar case, counts of violating lines are presented in
the error message.
In the latter case, only the file name can be presented due
to the "-zo" option to enable multi-line pattern search.

Note: There can be false negatives in the check.
For example, it can't not detect:

... \cref{label E}% some comment
{label F}

, at the moment.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>