treewide: Renew format of Quick Quiz macro
The format of \QuickQuiz{} macro before this change was not suitable
to be omitted entirely as suggested in the Answer to Quick Quiz 1.3.

For the question part to be ignored, it must be given as the argument
of the macro.

Update extractqqz.sh and qqz.sty accordingly.

Note that in debugging/debugging.tex, a verbatim snippet of VerbatimU
environment can not be held in \QuickQuiz's argument.

Instead, use fancyvrb's commands to save a verbatim content outside
the argument and to use the saved content inside the argument.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>