Use 'Arm' as text trademark of Arm architecture
Substitute "Arm" for "ARM" to respect the decision of Arm Limited.
Instead of direct substitutions, define macros \ARM and \ARMv.
This should help us easily catch up in case Arm changes its mind.

Note that as far as an argument of a macro is a single digit,
enclosing it by "{}" is not necessary.
For example, \ARMv{8} CPU" and "\ARMv8 CPU" will generate the same
result: "Armv8".

Some of "ARM" in ppcmem.tex are kept unchanged as the PPCMEM site
stilluses "ARM" as its interface choice.

Also update the legal page to mention trademarks of Arm, MIPS,
and SPARC.  Update the notice on Intel trademarks as well.
"x386" is not a trademark of Intel anymore.

While we are here, get rid of \mytexttrademark and \mytextregistered
as they have been empty ever since they were introduced in commit
eecdeac7367c ("Remove trademark and registered symbols in text").

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>