toolsoftrade: Improve rwlockscale.sh sampling strategy
The commit collects more samples at each point and collects the multiple
samples for a given point at widely separated times.  It also samples
fewer points for larger numbers of CPUs,reflecting the fact that the
curves all flatten out at large number of CPUs.  This approach avoids a
given disturbance on the system under test from affecting all samples
for a given point.  Note that the number of samples collected at each
point is insufficient for the normal average/standard-deviation approach,
but that is OK given that we instead plot maxima and minima.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>