utilities/torelease: Make sure new tag is pushed
Push the new tag to silence a warning from "git request-pull"
saying that the matching commit could not be found in the remote
repository.

Use variables "repo_url" and "remote" to make it possible to
use an alternative repository for contingency.

Also fix a grep pattern of tag existence check so that the
variable ${gittag} is evaluated.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>