appendix/whymb: Use 'step' as ref name in QQs
The \cref{} command added in commit e68dcb9a4f1a ("appendix/whymb:
Add QQ on cachelines in dirty shared state") puts "item" as the
reference name.  This is the default behavior for enumerate list.

As is observed in a couple of existing QQs in this appendix, "step"
is the right word to reference an entry in a sequence of operations.

With the help of "enumitem" package, we can define a new list
environment "sequence" and specify "step" as the name used by
cleveref's \cref{} commands.

Converting existing lists of sequences to the new "sequence" lists
and putting necessary labels can automatically use "step" in the
references to the sequence lists.

As a bonus, cross-references from the other QQs are now enabled.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>