toyrcu: Minor fixes
The floating snippet at the beginning of QQZ answer caused small skip
at the start of the answer part.
It might also cause an orphaned "Answer" heading.
Fix this by moving the snippet a few sentences down.

Note: I made a paragraph break in the middle of the long paragraph.

The other change is to use \O{} macro in the same way as the other
places.
I missed this one in commit 5e9a9c723a0b ("toyrcu: Use mathcal O for
'orders of'")

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>