utilities/autodate.sh: Make newer tcolorbox behave as before
tcolorbox 4.40 or later has wider skips around its boxes by default.

Putting \tcbsetforeverylayer{autoparskip} in preamble restores the
previous default behavior [1].

It needs to be put conditionally because the command changes older
tcolorbox's behavior.

We use a lot of booleans for conditional things in perfbook, but
their values are changed based on targets such as -1c, -nq, -qq, etc.
However, version of tcolorbox is common to all the targets.
We need a different way to put the \tcbsetforeverylayer{} command.

As autodate.tex is included in the preamble of perfbook, we can do
this by adding some script in autodate.sh, which will generate the
command when a newer tcolorbox is found.

Also update Makefile to let autodate.tex depend on autodate.sh.

[1]: https://www.ctan.org/ctan-ann/id/mailman.3088.1601153860.2548.ctan-ann@ctan.org

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>