datastruct/hash: Adjust for inappropriateness of global locking
Testing a globally locked hash table at 448 CPUs is pointless.  This
commit therefore adjusts the upper thread-count limits accordingly.
The data for the graph showing per-bucket locking converging with global
locking for a cat-only workload is collected at 64 CPUs.  Unless there
are fewer than 64 CPUs, in which case the actual number of CPUs is
used instead.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>