datastruct/hash: Correctly restrain hash_global update CPUs
We want to keep hash_global's number of CPUs restricted because runs
with large numbers of CPUs can take a long time and provide little or no
useful data.  However, while the F-series tests do restrict the number
of update-side CPUs, they cleverly set the number of readers to the
remaining number of CPUs out of the total.  This commit therefore adds
another check for hash_global and in that case instead sets the number
of readers to the amount remaining from smallncpus, thus limiting the
total number of threads to smallncpus.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>