datastruct/hash: Speed up hashtorture initial insertions
The current code scans for an uninserted item, which is good and
sufficient for hash tables with a few thousand buckets and items,
but quite slow for hash tables having millions of buckets and items.
This commit therefore uses a linear scan to do the insertion, with
a probabilistic check subject to bounds for each item that might be
inserted.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>