datastruct/hash: Unregister control thread
The zoo_test() does a bit of smoke testing, for which it must register
itself as an RCU reader.  However, this exposes it to signals, which
can cause the test to end too soon.  This commit avoids this problem by
unregistering the thread executing zoo_test() before the timed waits.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>