datastruct/hash: Data from perf.sh that is even less bogus
Remaining bogosity sources include lack of unsynchronized data, as in
running read-only tests with no synchronization whatsoever.  This version
avoids bottlenecks in glibc random() and fixes several bugs, as described
in earlier commits, but with the addition of a fix for a userspace-RCU
hang.  This hang was exposed by the earlier speedups.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>