CodeSamples/formal/litmus: Remove redundant initialization
In C Flavor litmus tests, global variables of type "int" need
initialization only when they aren't 0.
This is true not only for herd7 but also for klitmus7/litmus7.

Declaration of plain "int" or "int *" can also be removed.

We can also remove "&" in the right hand side of initialization.

Finally, to reduce line counts of snippets, use "{}" when
init blocks are empty.

For consistency, put an empty line above initialization blocks and
remove extra blank lines in other parts of litmus tests.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>